+++ Assets/Script/ScriptableObject/ArmySO.cs
@@ -103,7 +103,7 @@ public class ArmySO : ScriptableObject, IBaseInfo, IArmy, IWeapon, ISkill
         this.buildFacilities = buildFacilities;
     }
 
-    public void SetWeapon(string weaponNameZH, string weaponNameEN, DamageTypeSO damageType, Vector2 singleDamage, Vector2 range, float magazineSize, Vector2 magazineLoadingTime, Vector2 aimingTime, Vector2 firingDuration, Vector2 sputteringRadius, Vector2 sputteringDamage)
+    public void SetWeapon(string weaponNameZH, string weaponNameEN, DamageTypeSO damageType, Vector2 singleDamage, Vector2 range, int magazineSize, Vector2 magazineLoadingTime, Vector2 aimingTime, Vector2 firingDuration, Vector2 sputteringRadius, Vector2 sputteringDamage)
     {
         if (weapons == null)
             weapons = new List<Weapon>();
